# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Badran Car Wash Supplies & Car Accessories Store API",
    "version": "1.0.0",
    "description": "Static OpenAPI snapshot for the Badran Store modular monolith API."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register customer",
        "description": "Creates a customer account and returns the registered user profile.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Registration successful" },
          "400": { "description": "Validation failed or email already exists" }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login",
        "description": "Authenticates a registered user and returns a JWT bearer token.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "400": { "description": "Invalid email or password" }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get profile",
        "parameters": [{ "$ref": "#/components/parameters/UserIdHeader" }],
        "responses": {
          "200": { "description": "Profile retrieved" },
          "401": { "description": "JWT is missing or invalid" },
          "403": { "description": "Header user id does not match token user id" }
        }
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Request password reset",
        "parameters": [
          { "name": "email", "in": "query", "required": true, "schema": { "type": "string", "format": "email" }, "example": "ahmad.khalil@example.com" }
        ],
        "responses": {
          "200": { "description": "Password reset token generated" },
          "403": { "description": "Cannot request reset for another user" },
          "404": { "description": "Email does not exist" }
        }
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Reset password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ResetPasswordRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Password reset successful" },
          "400": { "description": "Invalid or expired reset token" }
        }
      }
    },
...
    "/api/v1/cart": {
      "get": {
        "tags": ["Cart"],
        "summary": "Get cart",
        "parameters": [{ "$ref": "#/components/parameters/OptionalUserIdHeader" }, { "$ref": "#/components/parameters/SessionIdHeader" }],
        "responses": { "200": { "description": "Cart retrieved" }, "400": { "description": "Missing user or session id" } }
      },
      "delete": {
        "tags": ["Cart"],
        "summary": "Clear cart",
        "parameters": [{ "$ref": "#/components/parameters/OptionalUserIdHeader" }, { "$ref": "#/components/parameters/SessionIdHeader" }],
        "responses": { "200": { "description": "Cart cleared" } }
      }
    },
    "/api/v1/cart/items": {
      "post": {
        "tags": ["Cart"],
        "summary": "Add cart item",
        "parameters": [{ "$ref": "#/components/parameters/OptionalUserIdHeader" }, { "$ref": "#/components/parameters/SessionIdHeader" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AddToCartRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Product added to cart" }, "400": { "description": "Validation failed" }, "404": { "description": "Product not found" } }
      }
    },
    "/api/v1/cart/items/{productId}": {
      "delete": {
        "tags": ["Cart"],
        "summary": "Remove cart item",
        "parameters": [{ "$ref": "#/components/parameters/OptionalUserIdHeader" }, { "$ref": "#/components/parameters/SessionIdHeader" }, { "$ref": "#/components/parameters/ProductIdPath" }],
        "responses": { "200": { "description": "Product removed from cart" }, "404": { "description": "Product not found in cart" } }
      }
    },
    "/api/v1/orders": {
      "post": {
        "tags": ["Orders"],
        "summary": "Create order",
        "parameters": [{ "$ref": "#/components/parameters/OptionalUserIdHeader" }, { "$ref": "#/components/parameters/SessionIdHeader" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateOrderRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Order created" }, "400": { "description": "Invalid checkout request" } }
      },
      "get": {
        "tags": ["Orders"],
        "summary": "List user orders",
        "parameters": [{ "$ref": "#/components/parameters/UserIdHeader" }],
        "responses": { "200": { "description": "Orders retrieved" } }
      }
    },
    "/api/v1/orders/{orderId}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get order",
        "parameters": [{ "$ref": "#/components/parameters/UserIdHeader" }, { "$ref": "#/components/parameters/OrderIdPath" }],
        "responses": { "200": { "description": "Order retrieved" }, "403": { "description": "Order does not belong to user" }, "404": { "description": "Order not found" } }
      }
    },
    "/api/v1/orders/{orderId}/payment": {
      "post": {
        "tags": ["Orders"],
        "summary": "Pay for order",
        "parameters": [
          { "$ref": "#/components/parameters/UserIdHeader" },
          { "$ref": "#/components/parameters/OrderIdPath" },
          { "name": "paymentMethod", "in": "query", "required": true, "schema": { "type": "string", "enum": ["cod", "card", "bank_transfer"] }, "example": "cod" }
        ],
        "responses": { "200": { "description": "Payment completed" }, "400": { "description": "Invalid payment request" }, "403": { "description": "Order does not belong to user" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "UserIdHeader": { "name": "X-User-Id", "in": "header", "required": true, "schema": { "type": "integer", "format": "int64" }, "example": 1 },
      "OptionalUserIdHeader": { "name": "X-User-Id", "in": "header", "required": false, "schema": { "type": "integer", "format": "int64" }, "example": 1 },
      "SessionIdHeader": { "name": "X-Session-Id", "in": "header", "required": false, "schema": { "type": "string" }, "example": "guest-session-1" },
      "ProductIdPath": { "name": "productId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" }, "example": 1 },
      "OrderIdPath": { "name": "orderId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" }, "example": 1 }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": { "type": "string", "example": "Ahmad Khalil" },
          "email": { "type": "string", "format": "email", "example": "ahmad.khalil@example.com" },
          "phone": { "type": "string", "example": "+970599123456" },
          "password": { "type": "string", "minLength": 6, "example": "password123" },
          "preferredLanguage": { "type": "string", "example": "en" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "ahmad.khalil@example.com" },
          "password": { "type": "string", "example": "password123" }
        }
      },
      "ResetPasswordRequest": {
        "type": "object",
        "required": ["token", "newPassword"],
        "properties": {
          "token": { "type": "string", "example": "550e8400-e29b-41d4-a716-446655440000" },
          "newPassword": { "type": "string", "minLength": 6, "example": "newPassword123" }
        }
      },
      "AddToCartRequest": {
        "type": "object",
        "required": ["productId", "quantity"],
        "properties": {
          "productId": { "type": "integer", "format": "int64", "example": 1 },
          "quantity": { "type": "integer", "minimum": 1, "example": 2 }
        }
      },
      "CreateOrderRequest": {
        "type": "object",
        "required": ["fulfillmentMethod", "paymentMethod"],
        "properties": {
          "fulfillmentMethod": { "type": "string", "example": "home_delivery" },
          "guestName": { "type": "string", "example": "Ahmad Khalil" },
          "guestPhone": { "type": "string", "example": "+970599123456" },
          "guestEmail": { "type": "string", "format": "email", "example": "ahmad.khalil@example.com" },
          "deliveryAddressId": { "type": "integer", "format": "int64", "example": 7 },
          "deliveryCity": { "type": "string", "example": "Ramallah" },
          "deliveryZone": { "type": "string", "example": "Al-Tireh" },
          "deliveryAddressLine": { "type": "string", "example": "Main Street, Building 12" },
          "couponCode": { "type": "string", "example": "WELCOME10" },
          "paymentMethod": { "type": "string", "example": "cod" }
        }
      }
    }
  }
}